LayoutTests/fast/dom/HTMLMeterElement/meter-clone-expected.txt

PASS cloned.value is target.value
PASS internals.userAgentPartId(clonedInnerElement) is internals.userAgentPartId(targetInnerElement)
PASS internals.userAgentPartId(clonedInnerElement.firstChild) is internals.userAgentPartId(targetInnerElement.firstChild)
PASS internals.userAgentPartId(clonedInnerElement.firstChild.firstChild) is internals.userAgentPartId(targetInnerElement.firstChild.firstChild)
PASS clonedInnerElement.firstChild.firstChild.style.inlineSize is "70%"
PASS targetInnerElement.firstChild.firstChild.style.inlineSize is "50%"
PASS successfullyParsed is true

TEST COMPLETE
